Refactor filenames for easier PR comparison
